+++ src/core/modules/instance/controllers/reactive.controller.ts
@@ -1,0 +1,20 @@
+import { reactive } from 'vue'
+import { InstanceService } from '../instance.service'
+import { AnyObject } from '../../../types/helper.types'
+import { recursiveReassign } from '../../../utils/object'
+
+export class ReactiveController {
+  InstanceService: InstanceService
+  InitialReactiveState: AnyObject
+  ReactiveState: AnyObject
+
+  constructor(InstanceService: InstanceService) {
+    this.InstanceService = InstanceService
+    this.InitialReactiveState = structuredClone(this.InstanceService.machineSchema?.reactive ?? {})
+    this.ReactiveState = reactive(this.InstanceService.machineSchema?.reactive ?? {})
+  }
+
+  resetReactive = () => {
+    recursiveReassign(this.ReactiveState, this.InitialReactiveState)
+  }
+}